requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
psycopg2-binary 
pydantic
strictyaml